Further refactor
